heart/requirements.txt
Flask
Pillow
aggdraw
numpy